Convert tests to typescript
