+++ package.json
@@ -39,10 +39,19 @@
     "spdy": "^3.3.3"
   },
   "devDependencies": {
+    "@types/assert": "0.0.29",
+    "@types/chai": "^3.4.34",
+    "@types/chai-as-promised": "0.0.29",
+    "@types/mocha": "^2.2.33",
+    "@types/sinon": "^1.16.31",
+    "@types/supertest": "^1.1.31",
+    "@types/tmp": "0.0.31",
     "chai": "^3.5.0",
     "chai-as-promised": "^6.0.0",
     "gulp": "^3.9.1",
     "gulp-mocha": "^2.2.0",
+    "gulp-spawn-mocha": "^3.1.0",
+    "istanbul": "^0.4.5",
     "mocha": "^3.1.0",
     "sinon": "^1.17.5",
     "supertest": "^1.2.0",
